backend-microservices/ai-service-fastapi/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
easyocr==1.7.2
exceptiongroup==1.3.1
fastapi==0.134.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.25
tomli==2.4.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.18
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
